+++ tests/integration/deprecated/readerNotes-filter.test.js
@@ -11,8 +11,8 @@ const {
   createDocument,
   createTag,
   addNoteToCollection
-} = require('../utils/testUtils')
-const { urlToId } = require('../../utils/utils')
+} = require('../../utils/testUtils')
+const { urlToId } = require('../../../utils/utils')
 
 const test = async app => {
   const token = getToken()
