+++ tests/integration/deprecated/readerNotes-paginate.test.js
@@ -8,9 +8,9 @@ const {
   createPublication,
   createNote,
   createDocument
-} = require('../utils/testUtils')
+} = require('../../utils/testUtils')
 
-const { urlToId } = require('../../utils/utils')
+const { urlToId } = require('../../../utils/utils')
 
 const test = async app => {
   const token = getToken()
